refactor: scv.Run() as main routine blocker
