+++ pkg/exporter/exporter.go
@@ -1,0 +1,277 @@
+//go:build windows
+
+package exporter
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	_ "net/http/pprof"
+	"os"
+	"os/signal"
+	"os/user"
+	"runtime"
+	"strings"
+	"syscall"
+
+	winlog "github.com/prometheus-community/windows_exporter/pkg/log"
+	"github.com/prometheus-community/windows_exporter/pkg/types"
+	"github.com/prometheus-community/windows_exporter/pkg/utils"
+	"github.com/prometheus-community/windows_exporter/pkg/wmi"
+
+	"github.com/alecthomas/kingpin/v2"
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/prometheus-community/windows_exporter/pkg/collector"
+	"github.com/prometheus-community/windows_exporter/pkg/config"
+	"github.com/prometheus-community/windows_exporter/pkg/log/flag"
+	"github.com/prometheus/common/version"
+	"github.com/prometheus/exporter-toolkit/web"
+	webflag "github.com/prometheus/exporter-toolkit/web/kingpinflag"
+)
+
+var (
+	app        = kingpin.New("windows_exporter", "A metrics collector for Windows.")
+	configFile = app.Flag(
+		"config.file",
+		"YAML configuration file to use. Values set in this file will be overridden by CLI flags.",
+	).String()
+	insecure_skip_verify = app.Flag(
+		"config.file.insecure-skip-verify",
+		"Skip TLS verification in loading YAML configuration.",
+	).Default("false").Bool()
+	metricsPath = app.Flag(
+		"telemetry.path",
+		"URL path for surfacing collected metrics.",
+	).Default("/metrics").String()
+	disableExporterMetrics = app.Flag(
+		"web.disable-exporter-metrics",
+		"Exclude metrics about the exporter itself (promhttp_*, process_*, go_*).",
+	).Bool()
+	maxRequests = app.Flag(
+		"telemetry.max-requests",
+		"Maximum number of concurrent requests. 0 to disable.",
+	).Default("5").Int()
+	enabledCollectors = app.Flag(
+		"collectors.enabled",
+		"Comma-separated list of collectors to use. Use '[defaults]' as a placeholder for all the collectors enabled by default.").
+		Default(types.DefaultCollectors).String()
+	printCollectors = app.Flag(
+		"collectors.print",
+		"If true, print available collectors and exit.",
+	).Bool()
+	timeoutMargin = app.Flag(
+		"scrape.timeout-margin",
+		"Seconds to subtract from the timeout allowed by the client. Tune to allow for overhead or high loads.",
+	).Default("0.5").Float64()
+	webConfig = webflag.AddFlags(app, ":9182")
+)
+
+// Same struct prometheus uses for their /version endpoint.
+// Separate copy to avoid pulling all of prometheus as a dependency
+type prometheusVersion struct {
+	Version   string `json:"version"`
+	Revision  string `json:"revision"`
+	Branch    string `json:"branch"`
+	BuildUser string `json:"buildUser"`
+	BuildDate string `json:"buildDate"`
+	GoVersion string `json:"goVersion"`
+}
+
+type Exporter struct {
+	logger       log.Logger
+	winlogConfig winlog.Config
+	collectors   collector.Collectors
+}
+
+func New() *Exporter {
+	e := &Exporter{}
+
+	e.winlogConfig = winlog.Config{}
+	flag.AddFlags(app, &e.winlogConfig)
+
+	app.Version(version.Print("windows_exporter"))
+	app.HelpFlag.Short('h')
+
+	// Initialize collectors before loading and parsing CLI arguments
+	e.collectors = collector.NewWithFlags(app)
+
+	e.loadConfiguration()
+	e.initiateWmi()
+	e.printRunAsUser()
+	return e
+}
+
+func (e *Exporter) GetLogger() log.Logger {
+	return e.logger
+}
+
+func (e *Exporter) PrintCollectors() bool {
+	return *printCollectors
+}
+
+func (e *Exporter) Start() {
+	e.buildCollectors()
+	e.buildAndStartHttpServer()
+}
+
+func (e *Exporter) RunAsCli() {
+	e.Start()
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, os.Interrupt, syscall.SIGTERM)
+	<-sigs
+	_ = level.Info(e.logger).Log("msg", "Shutting down windows_exporter")
+}
+
+func (e *Exporter) loadConfiguration() {
+	// Load values from configuration file(s). Executable flags must first be parsed, in order
+	// to load the specified file(s).
+	kingpin.MustParse(app.Parse(os.Args[1:]))
+
+	logger, err := winlog.New(&e.winlogConfig)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Failed to initialize logger: %v\n", err)
+		os.Exit(1)
+	}
+	e.logger = log.With(logger, "main", "exporter")
+	e.collectors.SetLogger(logger)
+
+	_ = level.Debug(e.logger).Log("msg", "Logging has Started")
+
+	if *configFile != "" {
+		resolver, err := config.NewResolver(*configFile, logger, *insecure_skip_verify)
+		if err != nil {
+			_ = level.Error(e.logger).Log("msg", "could not load config file", "err", err)
+			os.Exit(1)
+		}
+		err = resolver.Bind(app, os.Args[1:])
+		if err != nil {
+			_ = level.Error(e.logger).Log("err", err)
+			os.Exit(1)
+		}
+
+		// NOTE: This is temporary fix for issue #1092, calling kingpin.Parse
+		// twice makes slices flags duplicate its value, this clean up
+		// the first parse before the second call.
+		*webConfig.WebListenAddresses = (*webConfig.WebListenAddresses)[1:]
+
+		// Parse flags once more to include those discovered in configuration file(s).
+		kingpin.MustParse(app.Parse(os.Args[1:]))
+	}
+}
+
+func (e *Exporter) initiateWmi() {
+	if err := wmi.InitWbem(e.logger); err != nil {
+		_ = level.Error(e.logger).Log("Initiate SWbemServices failed with: ", "err", err)
+		os.Exit(1)
+	}
+}
+
+func (e *Exporter) printRunAsUser() {
+	if u, err := user.Current(); err != nil {
+		_ = level.Warn(e.logger).Log("msg", "Unable to determine which user is running this exporter. More info: https://github.com/golang/go/issues/37348")
+	} else {
+		_ = level.Info(e.logger).Log("msg", fmt.Sprintf("Running as %v", u.Username))
+
+		if strings.Contains(u.Username, "ContainerAdministrator") || strings.Contains(u.Username, "ContainerUser") {
+			_ = level.Warn(e.logger).Log("msg", "Running as a preconfigured Windows Container user. This may mean you do not have Windows HostProcess containers configured correctly and some functionality will not work as expected.")
+		}
+	}
+}
+
+func (e *Exporter) buildCollectors() {
+	enabledCollectorList := utils.ExpandEnabledCollectors(*enabledCollectors)
+	e.collectors.Enable(enabledCollectorList)
+	_ = level.Info(e.logger).Log("msg", fmt.Sprintf("Enabled collectors: %v", strings.Join(enabledCollectorList, ", ")))
+
+	// Initialize collectors before loading
+	err := e.collectors.Build()
+	if err != nil {
+		_ = level.Error(e.logger).Log("msg", "Couldn't load collectors", "err", err)
+		os.Exit(1)
+	}
+}
+
+func (e *Exporter) buildAndStartHttpServer() {
+	http.HandleFunc(*metricsPath, withConcurrencyLimit(*maxRequests, e.collectors.BuildServeHTTP(*disableExporterMetrics, *timeoutMargin)))
+	http.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		_, err := fmt.Fprintln(w, `{"status":"ok"}`)
+		if err != nil {
+			_ = level.Debug(e.logger).Log("Failed to write to stream", "err", err)
+		}
+	})
+	http.HandleFunc("/version", func(w http.ResponseWriter, r *http.Request) {
+		// we can't use "version" directly as it is a package, and not an object that
+		// can be serialized.
+		err := json.NewEncoder(w).Encode(prometheusVersion{
+			Version:   version.Version,
+			Revision:  version.Revision,
+			Branch:    version.Branch,
+			BuildUser: version.BuildUser,
+			BuildDate: version.BuildDate,
+			GoVersion: version.GoVersion,
+		})
+		if err != nil {
+			http.Error(w, fmt.Sprintf("error encoding JSON: %s", err), http.StatusInternalServerError)
+		}
+	})
+	if *metricsPath != "/" && *metricsPath != "" {
+		landingConfig := web.LandingConfig{
+			Name:        "Windows Exporter",
+			Description: "Prometheus Exporter for Windows servers",
+			Version:     version.Info(),
+			Links: []web.LandingLinks{
+				{
+					Address: *metricsPath,
+					Text:    "Metrics",
+				},
+				{
+					Address: "/health",
+					Text:    "Health Check",
+				},
+				{
+					Address: "/version",
+					Text:    "Version Info",
+				},
+			},
+		}
+		landingPage, err := web.NewLandingPage(landingConfig)
+		if err != nil {
+			_ = level.Error(e.logger).Log("msg", "failed to generate landing page", "err", err)
+			os.Exit(1)
+		}
+		http.Handle("/", landingPage)
+	}
+
+	_ = level.Info(e.logger).Log("msg", "Starting windows_exporter", "version", version.Info())
+	_ = level.Info(e.logger).Log("msg", "Build context", "build_context", version.BuildContext())
+	_ = level.Debug(e.logger).Log("msg", "Go MAXPROCS", "procs", runtime.GOMAXPROCS(0))
+
+	go func() {
+		server := &http.Server{}
+		if err := web.ListenAndServe(server, webConfig, e.logger); err != nil {
+			_ = level.Error(e.logger).Log("msg", "cannot start windows_exporter", "err", err)
+			os.Exit(1)
+		}
+	}()
+}
+
+func withConcurrencyLimit(n int, next http.HandlerFunc) http.HandlerFunc {
+	if n <= 0 {
+		return next
+	}
+
+	sem := make(chan struct{}, n)
+	return func(w http.ResponseWriter, r *http.Request) {
+		select {
+		case sem <- struct{}{}:
+			defer func() { <-sem }()
+		default:
+			w.WriteHeader(http.StatusServiceUnavailable)
+			_, _ = w.Write([]byte("Too many concurrent requests"))
+			return
+		}
+		next(w, r)
+	}
+}
